job control: set terminal group ID even for scripts (re: deb79dc3)
On scripts (including -c scripts) with -m/--monitor on, the
referenced commit introduced a nasty bug which results in various
full-screen terminal programs, such as top(1) and joe(1), hanging
or otherwise misbehaving on startup. Reproducers:

   $ ksh -m -c 'top; :'
   $ ksh -c 'set -m; top; :'

Analysis: The problem is that we're setting the process group ID
(using setpgid(2)) without also setting the terminal foreground
process group ID (using tcsetpgrp(2)). The latter is only done on
interactive shells. On scripts with -m on, we only call setpgid(2)
but that is not correct if the script's shell has a controlling
terminal. So we have to fully initialise jobcontrol and set the
job.jobcontrol flag even in that context.

src/cmd/ksh93/include/jobs.[ch]:
- Split job_init() function into a part for the non-tty bits of job
  control and the tty bits, which are now in job_init_tty(). Call
  this if from job_tty() if the shell is interactive. When it is
  called, set a new job.inited flag on successful completion.
- Instead of bailing out on most of job control init if the shell
  is not interactive, only disable the bits that can only work on
  interactive shells. This includes actively getting a controlling
  terminal if needed, waiting until the shell is in the foreground,
  and setting the SIGTSTP handler to sh_fault for it to handle ^Z.
- Fix a couple of uses of job.jobcontrol vs sh_isstate(SH_MONITOR).

src/cmd/ksh93/sh/args.c: applyopts():
- When turning on -m, call job_init_tty() if that has not yet been
  done. This ensures setting job.jobcontrol if possible.
- When toggling -m, toggle the entirety of job control: sync both
  the SH_MONITOR state bit and the job.jobcontrol flag when the -m
  option is toggled.

src/cmd/ksh93/sh/xec.c:
- check_exec_optimization(): Job control necessitates setting a new
  process group, which would not happen if we replace the shell by
  simply execve'ing the command. So never do the exec optimization
  when the SH_MONITOR state bit is on. (re: 88aa369e)

src/cmd/ksh93/sh/init.c:
- sh_reinit(): Reset job.inited flag when re-initing the shell to
  run a script; this ensures job_init_tty() will be called again
  from args.c when -m is turned on in the script.

+++ src/cmd/ksh93/sh/jobs.c
@@ -27,24 +27,26 @@
  *  Revised January, 1992
  *  Mended February, 2021
  *  Corrected March, 2025
+ *  Re-amended July, 2026
  *
  *  Aspects of job control are (de)activated using a global flag variable,
  *  a state bit, and a shell option bit. It is important to understand the
  *  difference and set/check them in a manner consistent with their purpose.
  *
- *  1. The job.jobcontrol flag is for job control on interactive shells.
- *     It is set to nonzero by job_init() if, and only if, the shell is
- *     interactive *and* managed to get control of the terminal. Therefore,
+ *  1. The job.jobcontrol flag is for the job control aspects to do with the
+ *     tty. It is set to nonzero by job_init_tty() if the shell managed to
+ *     get control of the terminal. This is done at init time on interactive
+ *     shells, or, for scripts, when the -m option is turned on. Therefore,
  *     any changing of terminal settings (tcsetpgrp(3), tty_set()) should
  *     only be done if job.jobcontrol is nonzero.
  *
  *  2. The state flag, sh_isstate(SH_MONITOR), determines whether the bits
- *     of job control that are relevant for both scripts and interactive
- *     shells are active, which is mostly making sure that each job gets
+ *     of job control that are relevant with and without control of the
+ *     terminal are active, which is mostly making sure that each job gets
  *     its own process group (setpgid(3)).
  *
  *  3. The -m (-o monitor) shell option, sh_isoption(SH_MONITOR), is just
- *     that. When the user turns it on or off, the state flag is synched
+ *     that. When the user turns it on or off, the flags above are synched
  *     with it. It should usually not be directly checked for, as the state
  *     may be temporarily turned off without turning off the option.
  */
@@ -498,16 +500,23 @@ static void job_waitsafe(int sig)
  */
 void job_init(void)
 {
-	int ntry=0;
 	job.fd = JOBTTY;
 	signal(SIGCHLD,job_waitsafe);
 	if(njob_savelist < NJOB_SAVELIST)
 		init_savelist();
-	if(!sh_isoption(SH_INTERACTIVE))
-		return;
 	job.mypgid = getpgrp();
+	/*
+	 * On interactive shells, initialize the tty aspects of job control now.
+	 * (For scripts, we do this from args.c when the -m option is turned on.)
+	 */
+	if(sh_isoption(SH_INTERACTIVE))
+		job_init_tty();
+}
+
+void job_init_tty(void)
+{
 	/* some systems have job control, but not initialized */
-	if(job.mypgid<=0)
+	if(sh_isoption(SH_INTERACTIVE) && job.mypgid<=0)
 	{
 		/* Get a controlling terminal and set process group */
 		/* This should have already been done by rlogin */
@@ -526,18 +535,22 @@ void job_init(void)
 	possible = (setpgid(0,job.mypgid) >= 0) || errno==EPERM;
 	if(possible)
 	{
+		int ntry = 0;
 		/* wait until we are in the foreground */
 		while((job.mytgid=tcgetpgrp(JOBTTY)) != job.mypgid)
 		{
 			if(job.mytgid <= 0)
 				return;
+			/* Avoid hang witout -i, or with -i -c */
+			if(!sh_isoption(SH_INTERACTIVE) || sh_isoption(SH_CFLAG))
+				break;
 			/* Stop this shell until continued */
 			signal(SIGTTIN,SIG_DFL);
 			kill(sh.pid,SIGTTIN);
 			/* resumes here after continue tries again */
 			if(ntry++ > IOMAXTRY)
 			{
-				errormsg(SH_DICT,0,e_no_start);
+				errormsg(SH_DICT,ERROR_warn(0),e_no_start);
 				return;
 			}
 		}
@@ -559,7 +572,8 @@ void job_init(void)
 	signal(SIGTTIN,SIG_IGN);
 	signal(SIGTTOU,SIG_IGN);
 	/* The shell now handles ^Z */
-	signal(SIGTSTP,sh_fault);
+	if(sh_isoption(SH_INTERACTIVE))
+		signal(SIGTSTP,sh_fault);
 	tcsetpgrp(JOBTTY,sh.pid);
 #ifdef CNSUSP
 	/* set the switch character */
@@ -571,8 +585,8 @@ void job_init(void)
 		tty_set(JOBTTY,TCSAFLUSH,&my_stty);
 	}
 #endif /* CNSUSP */
-	sh_onoption(SH_MONITOR);
-	job.jobcontrol++;
+	job.jobcontrol = 1;
+	job.inited = 1;
 	return;
 }
 
@@ -620,7 +634,7 @@ int job_close(void)
 		}
 	}
 	job_unlock();
-	if(job.jobcontrol && setpgid(0,job.mypgid)>=0)
+	if(setpgid(0,job.mypgid)>=0 && job.jobcontrol)
 		tcsetpgrp(job.fd,job.mypgid);
 #   ifdef CNSUSP
 	if(possible && job.suspend==CNSUSP)
@@ -906,7 +920,7 @@ int job_kill(struct process *pw,int sig)
 	pid = pw->p_pid;
 	if(by_number)
 	{
-		if(pid==0 && job.jobcontrol)
+		if(pid==0 && sh_isstate(SH_MONITOR))
 			r = job_walk(outfile, job_kill,sig, NULL);
 		if(sig==SIGSTOP && pid==sh.pid && sh_isoption(SH_LOGIN_SHELL))
 		{
